+++ vector.h
@@ -10,14 +10,14 @@ struct Vector
      void (*FreeNodeFn) (void*);
 };
 
-int Vector_Create(struct Vector* pArray, size_t initial_size, void (*FreeNodeFunc) (void*));
+size_t Vector_Create(struct Vector* pArray, size_t initial_size, void (*FreeNodeFunc) (void*));
 void Vector_Destroy(struct Vector* pArray);
 
 size_t Vector_Count(struct Vector* pArray);
-int Vector_Push(struct Vector* pArray, void* pVal);
+size_t Vector_Push(struct Vector* pArray, void* pVal);
 void* Vector_At(struct Vector* pArray, size_t idx);
-int Vector_Find(struct Vector* pArray, void* key,
+size_t Vector_Find(struct Vector* pArray, void* key,
 		int (*comp)(void*, void*), size_t* foundidx);
-int Vector_Remove(struct Vector* pArray, size_t idx);
+size_t Vector_Remove(struct Vector* pArray, size_t idx);
 
 #endif
